Move `StdTx`, `isStdTx` and `makeStdTx` to @cosmjs/amino
