+++ packages/amino/src/index.ts
@@ -31,4 +31,5 @@ export { Secp256k1Wallet } from "./secp256k1wallet";
 export { decodeSignature, encodeSecp256k1Signature, StdSignature } from "./signature";
 export { AminoMsg, makeSignDoc, serializeSignDoc, StdFee, StdSignDoc } from "./signdoc";
 export { AccountData, Algo, AminoSignResponse, OfflineAminoSigner } from "./signer";
+export { StdTx, isStdTx, makeStdTx } from "./stdtx";
 export { executeKdf, KdfConfiguration } from "./wallet";
